requirements.txt
pytest>=7.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
httpx>=0.25.0
